Convert webconsole/actions/index.js to TS
