+++ src/devtools/client/webconsole/actions/index.ts
@@ -1,0 +1,35 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+
+export { toggleTimestamps } from "devtools/client/webconsole/reducers/ui";
+export {
+  clearMessages,
+  logpointMessagesCleared,
+  messageClosed,
+  messageEvaluationsCleared,
+  messageOpened,
+  messagesAdded,
+  filterTextUpdated,
+  filterToggled,
+  exceptionLogpointErrorCleared,
+  exceptionLogpointErrorReceived,
+  setConsoleOverflowed,
+  setLastFetchedForFocusRegion,
+  setMessagesLoaded,
+} from "devtools/client/webconsole/reducers/messages";
+export * from "devtools/client/webconsole/actions/messages";
+export * from "devtools/client/webconsole/actions/toolbox";
+
+export {
+  evaluateExpression,
+  paywallExpression,
+  eagerEvalExpression,
+} from "devtools/client/webconsole/actions/input";
+
+export interface DebuggerLocation {
+  url?: string;
+  sourceId?: string;
+  line?: number;
+  column?: number;
+}
