+++ src/devtools/client/webconsole/components/FilterBar/FilterSearchBox.tsx
@@ -1,8 +1,7 @@
 import React, { ChangeEvent, KeyboardEvent, useState } from "react";
 import { useAppDispatch } from "ui/setup/hooks";
 import MaterialIcon from "ui/components/shared/MaterialIcon";
-// TODO Change require to import and convert the actions file to TS
-const actions = require("devtools/client/webconsole/actions/index");
+import * as actions from "devtools/client/webconsole/actions/index";
 
 function TextInput(props: React.HTMLProps<HTMLInputElement>) {
   return (
