+++ src/devtools/client/webconsole/utils/connected-smart-trace.tsx
@@ -4,4 +4,5 @@ import SmartTrace from "devtools/client/shared/components/SmartTrace";
 
 export default connect(null, {
   onViewSourceInDebugger: actions.onViewSourceInDebugger,
+  // @ts-expect-error some nested field mismatch
 })(SmartTrace);
